Reorganize epub package in a more traditionnal POO style
